make check_on_lilac_runs more informative
I don't think it was good that you couldn't see which LILAC run was which from that squeue call, so I found a way using grep to be just as selective while also showing more information, and more compactly (without compormising clarity!  the only annoying thing I've noticed is that it won't show jobs that ended more than an our or so ago.  but it seems like there's no way around that without using sacct, and I don't want to do that.  so this plus runs.log should be fine for now.