DO NOT pass an argument to lilac binary
apparently no matter what you pass it extremely screws up the output filenames.  I love the fact that this behavior is documented /s.  (tho really it just shouldn't exist... I tried and failed to find a way to use this feature in a way that doesn't ruin everything).  while I was at it I reduced the amount of output, because let's be real, we don't need this.