make test work consistently
the way I was testing the output, running the tests multiple times caused it to appear to fail.  so I added an option to write_row_to_outputs_table to remove all preexisting data, which gives the first test of that function a clean slate.  I'll also have use of this function later when I figure out how to log timeouts.