many improvements to output pdf
the thicknesses were printed rong, so that's fixd now.  the density is included in the composition slide instead of the material code.  DT is listed before DD.  instead of using the sums of DT+DD+D3He and D3He+ko-d to do bang-time stuff, it uses the maximum of them.  because who the heck sums reaction rates?  the yield integration is fixd to no longer return negative ko-d yields.  the burn width is listed in ps in the PDF.  it doesn't fail when there's no fusion (a rare occurrence, but when it does happen I would like output).