include example run_inputs.csv
I realized that I can have this file in the tree but also have it in the gitignore so that I don't spam the commit history with a bunch of simulation details.  truthfully I should have realized the implications when I first gitignored it and had to then igure out how to also remove it from the tree.  anyway, this way new users won't have to bild their own from scratch.