succeed where other scripts have failed
now the postprocessor will run after the simulation even if the simulation times out.  in addition, it will save the actual new state to run_ouputs.csv rather than blindly calling it "completed", and as a final touch, will exit peacefully if there is no output to process (for example if it timed out).