more intelligent run checking
it's always botherd me how squeue returns just a header when there are no runs.  I think it's nicer if it actually says "no runs".  so that's what this is.  it was a little tricky just because of how grep doesn't put a newline on the end of its output because echo always puts a newline at the end of *its* output... but it works.