simplify import system
I always struggle with getting python imports to work correctly when multiple directories are involved.  it was somewhat easier than normal here since all my source files are in a single flat directory; the issue is that my tests were elsewhere.  I had been setting the pythonpath to the lettuce root every time I did a test or ran a script, but I realized it's better to let the pythonpath be lettuce/python most of the time *except* when testing.  so that's what I've done.  it took a while, but I figured out how to set the pythonpath before testing in both linux and windows, and removed everything messing with the pythonpath in cases where it's already in lettuce/python by default (which is the majority of cases).