show more output
if it exits because of "negative temperature detected" (I wish LILAC would just say "I did the math wrong and made a negative temperature".  take responsibility, my dude.) it is perfectly capable of generating output anyway except for this one line.  so with this change, it actually produces the plots and stuff.