fixed D always throwing an error
yeah somehow I haven't noticed for months that using deuterium in the gas fill causes an automatic error.  like I know I haven't been using this regularly but still.  I adjusted the tests to catch this sort of thing and made it so that natural elemental mass numbers are more explicitly expressed, both internally in the code (so errors are more likely to get thrown if it mixes up natural carbon with carbon-12) and in the output files.  I still have to do the full end-to-end test, so I really hope this works.