versioning struggles
I hit this annoying issue with an older version of pandas where there used to be a bug where it would ignore the dtype argument for the index column.  like, I can see how this is *only* an issue if you have a string collum that *might* only have numeric characters in it, and only if that collumn just happens to be the index column.  but still, this is an extremely annoying bug.  whatever, hopefully I'll get to use a newer version of pandas soon.  until then, this test doesn't even catch it, really.  I also added a complaint to the readme about the installed version of h5py on this module being incompatible with the installed version of numpy (tho I'm tempted to agree with the implied argument on numpy's documentation when they passive aggressively state that the feature h5py was trying to use had been deprecated for over 14 years).