fix IRIS (mostly)
so there are a few things going on in this commit.  the top-level summary is that I got rid of my IRIS input deck preparation stuff because it didn't work and now I use Lotus to do it instead, and then I also added that postprocessing step I always wanted to add.  part of this was generalizing the status given in run_outputs.csv.  now instead of just saying, for example, "running" or "completed", it says "running IRIS" or "completed LILAC".

the postprocessed PDF looks pretty okay, but there is a bug where the neutron image is all zeros in this test case I've run.  something's fishy about the input deck, I think.

also this is only working for stopping power model 1 rite now.  I really need to merge the lilac and iris folders and then make the model1/ and model2/ stuff be part of the existing nested directory structure.