remove broken test
I'm not actually sure what this dtype is when it works correctly; some kind of pandas "object" dtype.  but since this test doesn't catch the edge case that inspired it anyway (since the canonical run_inputs.csv has nonnumeric characters in its shot collum) and I also didn't fix that edge case (since it will only come up when using an old version of Pandas) I think it's easiest to just trash it.