simplify directory structure
I got rid of the "lilac" and "iris" subdirectories in each folder because it was unnecessary and it didn't mesh well with the fact that iris directories often need more specificity than lilac directories.  all it required was thinking about the filenames and making sure they were descriptive enuff for the lilac and iris stuff to coexist in one folder.  and hey, now specifying the IRIS stopping power model works!